refactor: seperate models
