+++ universelist/models/__init__.py
@@ -1,0 +1,2 @@
+from .bot import *
+from .user import *
